# internal/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: fullWA API
  version: 0.2.4-phase2
  summary: Multi-tenant customer communication platform API.
  description: |
    Every request is scoped to an organization determined from the
    authenticated principal (session cookie or API key). Errors follow
    RFC 7807 (`application/problem+json`).
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: local dev
...
paths:
...
  /api/v1/messages:
    post:
      tags: [messages]
      operationId: postMessagesSend
      summary: Enqueue an outbound message.
      description: |
        Persists the message as `queued`, enqueues a job on the
        `message.send` lane, and returns 202. The provider adapter is
        invoked asynchronously by the send worker; status transitions
        (`sent`, `delivered`, `read`, `failed`) arrive later via
        `MessageSent` / `MessageDelivered` / `MessageRead` / `MessageFailed`
        events and provider status webhooks.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendMessageRequest' }
      responses:
        '202':
          description: Message queued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SendMessageAccepted' }
        '400':
          description: Invalid request body.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401':
          description: Not authenticated.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: CSRF failure or missing permission.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: Conversation not found.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '424':
          description: The conversation has no configured integration.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /api/v1/conversations:
    get:
      tags: [conversations]
      operationId: getConversations
      summary: List conversations for the caller's org.
      description: |
        Phase 1 placeholder — returns an empty list. Replaced with the full
        list implementation in Phase 1 Task 4.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: A (possibly empty) page of conversations.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConversationList' }
        '401':
          description: Not authenticated.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/messages/{id}/read:
    post:
      tags: [messages]
      operationId: postMessageMarkRead
      summary: Mark a single inbound message as read to the provider.
      description: |
        Calls the channel adapter's mark-as-read API (Meta shows the blue
        double-tick on the customer's phone) and stamps `read_at` locally.
        Idempotent — outbound messages, messages without a
        `provider_message_id`, and messages already stamped as read return
        204 without side-effects.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
          description: Message ULID.
      responses:
        '204':
          description: Mark-as-read accepted.
        '401':
          description: Not authenticated.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: CSRF failure.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          description: Message not found.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '424':
          description: The message's integration cannot be resolved.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '502':
          description: The upstream provider rejected the mark-as-read call.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/conversations/{id}/read:
    post:
      tags: [conversations]
      operationId: postConversationMarkRead
      summary: Mark the newest unread inbound messages in a conversation as read.
      description: |
        Batch variant of `POST /api/v1/messages/{id}/read`. Iterates over
        the newest 50 messages in the conversation and calls the provider's
        mark-as-read for every inbound row with a `provider_message_id`
        that is not already stamped `read_at`. Partial failures are
        recorded server-side; the caller can retry.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
          description: Conversation ULID.
      responses:
        '204':
          description: Batch mark-as-read accepted.
        '401':
          description: Not authenticated.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: CSRF failure.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '502':
          description: Upstream provider rejected one or more calls.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/conversations/{id}/messages:
    get:
      tags: [messages]
      operationId: getConversationMessages
      summary: List messages in a conversation, newest first.
      description: |
        Cursor-based pagination. The `next_cursor` returned in the response
        should be echoed back as the `cursor` query parameter on the next
        page. Absent `next_cursor` means the caller reached the end.
      security:
        - sessionCookie: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
          description: Conversation ULID.
        - in: query
          name: cursor
          required: false
          schema: { type: string }
          description: Opaque cursor returned by a prior page.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: A page of messages.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageList' }
        '400':
          description: Bad request.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401':
          description: Not authenticated.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: fullwa_session
      description: Signed session cookie.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Programmatic access with per-org scoped key.
  schemas:
...
    SendMessageRequest:
      type: object
      required: [conversation_id, type]
      description: |
        Payload for POST /api/v1/messages. `type` selects the canonical
        message shape; the corresponding sub-object (`text`, `template`,
        `media`, `interactive`, `location`, `reaction`) must be populated.
        Everything is provider-neutral — the channel adapter interprets it.
      properties:
        conversation_id:
          type: string
          description: ULID of the target conversation.
        type:
          type: string
          enum: [text, template, image, video, audio, document, sticker, interactive, location, reaction]
        text:
          type: object
          description: Present when type=text.
          required: [body]
          properties:
            body: { type: string }
            preview_url: { type: boolean }
        template:
          type: object
          description: Present when type=template.
          required: [name, language]
          properties:
            name: { type: string }
            language: { type: string, description: 'BCP47 or Meta locale.' }
            namespace: { type: string }
            components:
              type: array
              items:
                type: object
        media:
          type: object
          description: Present when type is one of image/video/audio/document/sticker.
          properties:
            kind: { type: string }
            media_id: { type: string }
            url: { type: string, format: uri }
            mime_type: { type: string }
            caption: { type: string }
            file_name: { type: string }
        interactive:
          type: object
          description: Present when type=interactive; forwarded verbatim.
        location:
          type: object
          description: Present when type=location.
          required: [latitude, longitude]
          properties:
            latitude: { type: number, format: double }
            longitude: { type: number, format: double }
            name: { type: string }
            address: { type: string }
        reaction:
          type: object
          description: Present when type=reaction.
          required: [message_id]
          properties:
            message_id: { type: string }
            emoji: { type: string }
        idempotency_key:
          type: string
          description: |
            Optional client-supplied key echoed to the provider (e.g. Meta's
            `biz_opaque_callback_data`). When absent the server uses the
            message ID.
...
    SendMessageAccepted:
      type: object
      required: [message_id, status]
      properties:
        message_id:
          type: string
          description: ULID of the newly-created message row.
        status:
          type: string
          enum: [queued]
    Message:
      type: object
      required:
        - id
        - conversation_id
        - session_id
        - contact_id
        - channel
        - provider
        - direction
        - type
        - status
        - sender
        - recipient
        - created_at
      properties:
        id: { type: string }
        conversation_id: { type: string }
        session_id: { type: string }
        contact_id: { type: string }
        channel: { type: string }
        provider: { type: string }
        direction:
          type: string
          enum: [inbound, outbound]
        type: { type: string }
        status:
          type: string
          enum: [queued, sent, delivered, read, failed]
        sender: { type: string }
        recipient: { type: string }
        provider_message_id: { type: string }
        text: { type: string }
        media_url:
          type: string
          format: uri
          description: |
            Same-origin URL to fetch the media bytes. When the ingress
            pipeline has downloaded the provider's copy this is
            `/api/v1/media/{key}`; otherwise it may be a short-lived
            provider-native URL (Meta CDN, etc.).
        content_type:
          type: string
          description: MIME type persisted alongside the media bytes.
        location:
          type: object
          description: |
            Populated when `type == "location"`. Latitude/longitude are
            always present; name/address/url are optional.
          properties:
            latitude: { type: number, format: double }
            longitude: { type: number, format: double }
            name: { type: string }
            address: { type: string }
            url: { type: string, format: uri }
        contacts:
          type: array
          description: Populated when `type == "contacts"`. One entry per shared vCard.
          items:
            type: object
            properties:
              name: { type: string }
              phones:
                type: array
                items: { type: string }
              emails:
                type: array
                items: { type: string, format: email }
        reaction:
          type: object
          description: |
            Populated when `type == "reaction"`. `message_id` is the
            provider-native id (e.g. wamid) of the reacted-to message.
            Empty `emoji` means the reaction was removed.
          properties:
            emoji: { type: string }
            message_id: { type: string }
        interactive:
          type: object
          description: |
            Populated when `type == "interactive"` (or `type == "button"` for
            template quick replies). Kind mirrors Meta's discriminator —
            `button_reply`, `list_reply`, `button`, `cta_url`, `flow`, …
          properties:
            kind: { type: string }
            id: { type: string }
            title: { type: string }
            description: { type: string }
        reply_to_provider_message_id:
          type: string
          description: |
            Provider-native id (e.g. wamid) of the message this one replies
            to. Frontends use it to attach reactions to the referenced
            bubble or render a "replying to" chip.
        created_at: { type: string, format: date-time }
        sent_at: { type: string, format: date-time }
        delivered_at: { type: string, format: date-time }
        read_at: { type: string, format: date-time }
    MessageList:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items: { $ref: '#/components/schemas/Message' }
        next_cursor:
          type: string
          description: Opaque cursor for the next page; absent when done.
    Conversation:
      type: object
      required: [id, status, created_at]
      properties:
        id: { type: string }
        session_id: { type: string }
        contact_id: { type: string }
        status:
          type: string
          enum: [open, pending, resolved, reopened]
        assigned_user_id: { type: string }
        assigned_team_id: { type: string }
        priority:
          type: string
          enum: [low, normal, high, urgent]
        unread_count: { type: integer }
        last_message_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    ConversationList:
      type: object
      required: [conversations]
      properties:
        conversations:
          type: array
          items: { $ref: '#/components/schemas/Conversation' }
        next_cursor:
          type: string
          description: Opaque cursor for the next page; absent when done.
...
    Problem:
      type: object
      description: RFC 7807 problem details.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type.
          examples: ['https://fullwa.dev/errors/validation']
        title:
          type: string
          examples: ['validation failed']
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
          examples: [400]
        detail:
          type: string
        instance:
          type: string
          format: uri
        request_id:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
security:
  - sessionCookie: []
  - apiKey: []